% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointDichotomization.R
\name{Joint}
\alias{Joint}
\title{Compute Optimal Joint Cutpoints}
\arguments{
\item{mydata}{A data frame containing variables X1, X2, and Y.}

\item{py}{Prevalence of Y in the population (default = 0.5).}
}
\value{
A data frame with the optimal cutpoints for each classification metric.
}
\description{
Identifies optimal cutpoints for joint dichotomization of two continuous predictors.
to classify a binary outcome using different classification metrics.
}
\examples{
data <- data.frame(X1 = rnorm(500), X2 = rnorm(500), Y = sample(0:1, 500, replace = TRUE))
Joint(data, py = 0.5)
}
